+++ lindera/src/tokenizer.rs
@@ -1,4 +1,4 @@
-use std::{fmt, path::PathBuf};
+use std::fmt;
 
 use serde::{
     de::{self, MapAccess, SeqAccess, Visitor},
@@ -8,39 +8,12 @@ use serde::{
 use lindera_core::{
     dictionary::Dictionary, token::Token, user_dictionary::UserDictionary, viterbi::Lattice,
 };
-use lindera_dictionary::DictionaryKind;
 
 use crate::{
-    builder::{load_dictionary, load_user_dictionary},
+    dictionary::{load_dictionary, load_user_dictionary, DictionaryConfig, UserDictionaryConfig},
     mode::Mode,
     LinderaResult,
 };
-
-/// Dictionary config
-///
-/// Use this if you want to use a dictionary when tokenizing.
-///
-/// Either `kind` or `path` must be specified.
-///
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-pub struct DictionaryConfig {
-    /// Specify the kind of dictionary (IPADIC, UniDic, ko-dic, CC-CEDICT) if a self-contained dictionary is used for tokenization.
-    pub kind: Option<DictionaryKind>,
-    /// Specifies the path to a pre-built external dictionary if one is used.
-    pub path: Option<PathBuf>,
-}
-
-/// User dictionary config
-///
-/// Use this if you want to use a user dictionary when tokenizing.
-///
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
-pub struct UserDictionaryConfig {
-    /// Path to the user dictionary file.
-    pub path: PathBuf,
-    /// If the user dictionary was in CSV format, specify the dictionary type (IPADIC, UniDic, ko-dic or CC-CEDICT).
-    pub kind: Option<DictionaryKind>,
-}
 
 // Only the value specified by the feature flag is stored.
 pub const CONTAINED_DICTIONARIES: &[&str] = &[